requirements_tfidf.txt
# Minimal requirements for TF-IDF based analysis with name standardization
# This avoids heavy dependencies like sentence-transformers
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0
nltk>=3.6.0
openai>=1.0.0
python-dotenv>=0.19.0